requirements.txt
msal
requests
python-dotenv
azure-identity
azure-storage-blob
datetime
